Add files via upload
This is a Bot I've built using selenium on python. The bot uses a firefox browser to acess your own account and like tweets on a specific subject (such as python, javascript, burguer, etc). Bots using selenium and web browsers used to work better, but the constant changes to the html of websites such as twitter, facebook and instagram have made it difficult to track the elements of the page you want to interact with. 
PS: the bot wont like all tweets, since some of them are under differents tags and xpaths 
PS2: the bot has a loop of 10 scrolls so it wont go on forever and be flagged as a bot by tweeter